parce path of texture and color
i forget to check C color

+++ strings/get_next_line_utils.c
@@ -6,13 +6,13 @@
 /*   By: del-yaag <del-yaag@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/02 23:08:29 by del-yaag          #+#    #+#             */
-/*   Updated: 2023/08/06 16:42:30 by del-yaag         ###   ########.fr       */
+/*   Updated: 2023/08/18 20:43:36 by del-yaag         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../cub3d.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
 	int	i;
